+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog Fab-manager
 
+- Fix a bug: installation without nginx does not remove the service from the docker-compose file
+
 ## v4.3.3 2020 April 1st
 
 - Docker build will no longer embed development dependencies